Move builtin_functions to separate file
