+++ numbas/src/jme/mod.rs
@@ -1,2 +1,3 @@
-pub mod parser;
 pub mod ast;
+pub mod builtin_functions;
+pub mod parser;
